+++ CHANGELOG.md
@@ -4,6 +4,7 @@ vNext Release notes (TBD)
 * None
 
 ### Enhancements
+* Accessing `Realm.Sync` when sync is not enabled will no longer throw, but return `undefined`.
 * Better error messages when creating objects.
 * Added bundled TypeScript declarations of the Realm API.